+++ HISTORY.rst
@@ -6,7 +6,7 @@
 .. ++++++++++++++++++
 
 .. **Features and Improvements**
-* Project proposal change order date to create_date
+
 .. **Bugfixes**
 
 .. **Build**
@@ -21,6 +21,7 @@ latest (unreleased)
 
 **Features and Improvements**
 * Add membership end date on account view
+* Project proposal change order date to create_date
 
 **Bugfixes**